Add `sparse` mode to cover tool
This adds a new flag `sparse`. When set, the cover tool will only
instrument units that cannot be inferred from others. As an example if
the code looks like this

```
[unit1]
if <condition> {
        [unit3]
}
[unit2]
```
only the units 2 and 3 will be instrumented, because knowing the state
of both we can infer the state of unit 1.

This needs changes to the way the data is recovered, because after the
execution of the instrumented application the information would be
incomplete.