Add restorer for sparse mode data
This adds a new type `restorer`, in addition to two new flags `restore`
and `src`. The new type will be in charge to recover information for the
debugdump mode, when sparse mode is used in the `cover` tool.

The implementation mimics the process that the cover tool does
generating units out ouf the source file, and also recovering the
relationship between them, this is, whether a unit can be inferred from
other.

This new mode currently supports recovering information for programs made
out of `if` and `for` statements.